Analyticity-backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pytz
pydantic
geopandas
geopy
scipy
numpy
pandas